requirements.txt
Django==4.2.8
celery==5.3.6
google-api-core==2.16.2
python-dotenv==0.19.2
django-jazzmin==2.6.0
django-el-pagination==4.0.0
django_extensions==3.2.3
django-tailwind==3.6.0
django-debug-toolbar==4.2.0
django-cors-headers==4.3.1
django-storages==1.14.2
psycopg2==2.9.9
drf-yasg==1.21.7
pymemcache==4.0.0
django-grappelli==3.0.8
pydub==0.25.1
plone.rfc822==3.0.0
requests==2.26.0
requests-oauthlib==1.3.1
openai==1.11.1
gnews==0.3.6
langchain==0.1.5
azure-cognitiveservices-speech==1.35.0
gunicorn==21.2.0
django-storages==1.14.2
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
boto3==1.34.38
elevenlabs==1.2.0
